refactor(images): 👷 meta data and folders
