+++ lib/data/metadata/pages/projects/personal/metaDataKrsiak.ts
@@ -6,8 +6,8 @@ import { ENV_URLS, PAGES_URL } from '@/lib/utils/constants/urls/pageUrls'
 
 import { formatKeywordsString } from '@/lib/utils/helpers/metadata/formatKeywordsString'
 
-import krsiakOG from '@/public/images/png/open-graph/pages/personal-projects/krsiak-og.png'
-import krsiakTwitter from '@/public/images/png/open-graph/pages/personal-projects/krsiak-twitter.png'
+import krsiakOG from '@/public/images/png/open-graph/pages/personal-projects/krsiak/krsiak-og.png'
+import krsiakTwitter from '@/public/images/png/open-graph/pages/personal-projects/krsiak/krsiak-twitter.png'
 
 export const metaDataKrsiak: Metadata = {
   title: META_KRSIAK.title,
