+++ public/images/png/open-graph/pages/personal-projects/cryptomania/cryptomania-og.png
